requirements.txt
playwright
flask
flask-cors
beautifulsoup4
google-generativeai 